feat(sign): use signflinger to sign archive instead of apksigner
- Deprecated zipalign, because the archive signed with signflinger has
been aligned
- bump aapt2 to v34.0.3
- bump app version to 1.6

Closes #27
