+++ app/src/main/java/com/shapun/apkaabconverter/convert/AABToApkConverter.java
@@ -2,11 +2,10 @@ package com.shapun.apkaabconverter.convert;
 
 import android.content.Context;
 
-import com.android.apksig.ApkSigner;
 import com.android.tools.build.bundletool.androidtools.Aapt2Command;
 import com.android.tools.build.bundletool.commands.BuildApksCommand;
 import com.android.tools.build.bundletool.model.SigningConfiguration;
-import com.google.common.collect.ImmutableList;
+import com.shapun.apkaabconverter.signing.model.CertificateInfo;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
@@ -16,12 +15,12 @@ import java.nio.file.Path;
 public class AABToApkConverter extends FileConverter {
 
     private final File AAPT2Binary;
-    private final ApkSigner.SignerConfig signerConfig;
+    private final CertificateInfo signingCertInfo;
 
     public AABToApkConverter(Builder builder) {
         super(builder);
         AAPT2Binary = new File(getContext().getApplicationInfo().nativeLibraryDir, "libaapt2.so");
-        signerConfig = builder.signerConfig;
+        signingCertInfo = builder.signingCertInfo;
     }
 
     public void start() {
@@ -35,10 +34,9 @@ public class AABToApkConverter extends FileConverter {
                 .setOverwriteOutput(true)
                 .setApkBuildMode(BuildApksCommand.ApkBuildMode.UNIVERSAL)
                 .setOutputPrintStream(new PrintStream(outputStream));
-        if (signerConfig != null)
+        if (signingCertInfo != null)
             builder.setSigningConfiguration(SigningConfiguration.builder()
-                    .setSignerConfig(signerConfig.getPrivateKey(),
-                            ImmutableList.copyOf(signerConfig.getCertificates()))
+                    .setSignerConfig(signingCertInfo.getKey(),signingCertInfo.getCertificate())
                     .build());
         builder.build().execute();
         addLog(outputStream.toString());
@@ -46,14 +44,14 @@ public class AABToApkConverter extends FileConverter {
     }
 
     public static class Builder extends FileConverter.Builder<Builder> {
-        private ApkSigner.SignerConfig signerConfig;
+        private CertificateInfo signingCertInfo;
 
         public Builder(Context context, Path aabPath, Path outputPath) {
             super(context, aabPath, outputPath);
         }
 
-        public Builder setSignerConfig(ApkSigner.SignerConfig signerConfig) {
-            this.signerConfig = signerConfig;
+        public Builder setSigningCertInfo(CertificateInfo signingCertInfo) {
+            this.signingCertInfo = signingCertInfo;
             return this;
         }
