--- app/src/main/java/com/shapun/apkaabconverter/jks/JavaKeyStore.java
@@ -1,14 +1,0 @@
-package com.shapun.apkaabconverter.jks;
-
-import org.bouncycastle.jce.provider.BouncyCastleProvider;
-
-import java.security.KeyStore;
-
-/**
- * Created by qingyu on 2023-03-23.
- */
-public class JavaKeyStore extends KeyStore {
-    public JavaKeyStore() {
-        super(new JavaKeyStoreSpi(), new BouncyCastleProvider(), "JKS");
-    }
-}
