--- app/src/main/java/com/shapun/apkaabconverter/jks/LoadKeystoreException.java
@@ -1,13 +1,0 @@
-package com.shapun.apkaabconverter.jks;
-
-import java.io.IOException;
-
-/** Thrown by JKS.engineLoad() for errors that occur after determining the keystore is actually a JKS keystore. */
-@SuppressWarnings("serial")
-public class LoadKeystoreException extends IOException {
-
-	public LoadKeystoreException(String message) {
-		super(message);
-	}
-
-}
