+++ app/src/main/java/com/shapun/apkaabconverter/signing/jks/LoadKeystoreException.java
@@ -1,0 +1,13 @@
+package com.shapun.apkaabconverter.signing.jks;
+
+import java.io.IOException;
+
+/** Thrown by JKS.engineLoad() for errors that occur after determining the keystore is actually a JKS keystore. */
+@SuppressWarnings("serial")
+public class LoadKeystoreException extends IOException {
+
+	public LoadKeystoreException(String message) {
+		super(message);
+	}
+
+}
